#42 refactor: organize commands into nodes and edges packages
