+++ platform/src/use_cases/cli/commands/nodes/create_node.py
