+++ platform/src/use_cases/cli/commands/nodes/delete_node.py
