+++ platform/src/use_cases/cli/commands/nodes/edit_node.py
